api/pyproject.toml
[tool.poetry]
name = "outhouse"
version = "0.1.0"
description = ""
authors = ["Jan Hein de Jong <janhein.dejong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
alembic = "^1.9.2"
pymysql = "^1.0.2"
loguru = "^0.6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^22.12.0"
mypy = "^0.991"
sqlalchemy-stubs = "^0.4"
autoflake = "^2.0.0"

[tool.mypy]
plugins = "sqlmypy"

[tool.poe.tasks]
format = { shell = "scripts/format.sh"}
test = { shell = "scripts/test.sh"}
serve = "uvicorn outhouse.main:app --port 8000 --reload"
build = { shell = "scripts/build.sh" }
publish = [
    { shell = "scripts/check-commit.sh" }, 
    { ref = "test" }, 
    { shell = "scripts/build-and-publish.sh" }    
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
